gpu: move the imports a kernel scope needs into their own unit
Two kinds of definition have to be brought in before a kernel can be
built from a loop: the body of a `module procedure`, which arrives as a
whole submodule translation unit read back from a `.smod` file together
with the modules it depends on, and a derived type, which has to be
copied into the kernel's own symbol table along with every type its
members name.

Pure code motion: the definitions are unchanged, only their home is.
