+++ src/libasr/CMakeLists.txt
@@ -90,6 +90,7 @@ set(SRC
     pass/gpu_offload_device_fn.cpp
     pass/gpu_offload_intrinsics.cpp
     pass/gpu_offload_array_ops.cpp
+    pass/gpu_offload_import.cpp
     pass/device_partition.cpp
     pass/device_launch_expand.cpp
     pass/gpu_memory_space.cpp
